refactor: decouple code generating icons from dependencies components-icons
